refactor: rename `snapshot_uploaders` module to `file_uploaders`
