+++ mithril-aggregator/src/dependency_injection/containers.rs
@@ -32,13 +32,13 @@ use crate::{
     },
     entities::AggregatorEpochSettings,
     event_store::{EventMessage, TransmitterService},
+    file_uploaders::SnapshotUploader,
     multi_signer::MultiSigner,
     services::{
         CertifierService, EpochService, MessageService, ProverService, SignedEntityService,
         StakeDistributionService, TransactionStore, UpkeepService,
     },
     signer_registerer::SignerRecorder,
-    snapshot_uploaders::SnapshotUploader,
     store::CertificatePendingStorer,
     EpochSettingsStorer, MetricsService, SignerRegisterer, SignerRegistrationRoundOpener,
     SingleSignatureAuthenticator, Snapshotter, VerificationKeyStorer,
