+++ mithril-aggregator/src/file_uploaders/dumb_snapshot_uploader.rs
