+++ mithril-aggregator/src/file_uploaders/snapshot_uploader.rs
