+++ mithril-aggregator/src/lib.rs
@@ -18,6 +18,7 @@ pub mod database;
 pub mod dependency_injection;
 pub mod entities;
 pub mod event_store;
+mod file_uploaders;
 mod http_server;
 mod message_adapters;
 pub mod metrics;
@@ -25,7 +26,6 @@ mod multi_signer;
 mod runtime;
 pub mod services;
 mod signer_registerer;
-mod snapshot_uploaders;
 mod snapshotter;
 mod store;
 mod tools;
@@ -38,6 +38,9 @@ pub use crate::configuration::{
 pub use crate::multi_signer::{MultiSigner, MultiSignerImpl};
 pub use commands::{CommandType, MainOpts};
 pub use dependency_injection::DependencyContainer;
+pub use file_uploaders::{
+    DumbSnapshotUploader, LocalSnapshotUploader, RemoteSnapshotUploader, SnapshotUploader,
+};
 pub use message_adapters::{FromRegisterSignerAdapter, ToCertificatePendingMessageAdapter};
 pub use metrics::*;
 pub use runtime::{
@@ -46,9 +49,6 @@ pub use runtime::{
 pub use signer_registerer::{
     MithrilSignerRegisterer, SignerRecorder, SignerRegisterer, SignerRegistrationError,
     SignerRegistrationRound, SignerRegistrationRoundOpener,
-};
-pub use snapshot_uploaders::{
-    DumbSnapshotUploader, LocalSnapshotUploader, RemoteSnapshotUploader, SnapshotUploader,
 };
 pub use snapshotter::{
     CompressedArchiveSnapshotter, DumbSnapshotter, SnapshotError, Snapshotter,
